+++ EnglishAppUITests/Info.plist
