--- IroIroIro/.gitignore
@@ -1,5 +1,0 @@
-*.DS_Store
-Pods/
-*.xcworkspace
-EnglishApp.xcodeproj/xcuserdata/*
-EnglishApp/YOLOv3.mlmodel
