--- IroIroIro/EnglishApp/QuestionViewController.swift
@@ -1,34 +1,0 @@
-//
-//  Question.swift
-//  ARKit3DTextSample
-//
-//  Created by 新垣結梨 on 2019/10/19.
-//  Copyright © 2019 SAPPOROWORKS. All rights reserved.
-//
-
-import Foundation
-import UIKit
-
-class QuestionViewController: UIViewController {
-    
-    @IBOutlet weak var question: UILabel! // お題用のラベル
-    
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        
-        self.setupUI()
-        
-        DispatchQueue.main.asyncAfter(deadline: .now() + 3.0) {
-            // 3秒後にARの画面に遷移
-            self.performSegue(withIdentifier: "toARViewController", sender: nil)
-        }
-    }
-    
-    // MARK: - UIを整える関数
-    private func setupUI() {
-        question.text = "DOG"
-        question.textColor = .purple
-        question.textAlignment = .center
-        question.font = UIFont.systemFont(ofSize: 100.0)
-    }
-}
