--- IroIroIro/README.md
@@ -1,14 +1,0 @@
-# いぇいご
-英単語を楽しく覚えたい小学４〜6年生を対象としたAR英単語探しアプリケーションです
-
-## エレベーターピッチ
-`英単語を学習したい`  
-`小学生`向けの  
-` いぇいご（仮）`という  
-`スマホアプリ`です。  
-これは、`お題（英単語）に合わせて周囲からものを探すこと`ができ,  
-`英単語帳`とは違って,  
-`実際にものを見ながら英語を学習`ができます。  
-
-## プロジェクト管理
-https://trello.com/b/BfYlpbdq/%E3%81%84%E3%81%87%E3%81%84%E3%81%94
